Update data to allow models to leverage initial condition after gap
We want the model to start with the initial condition of abundance for
the first observation following the regime shift gap. The simplest way
to do this is to keepp the gap and the first newmoon post regime shift
in data_train and fill the y values for the gap (but not the initial
condition newmoon) with NA's and then start the test set on the newmoon
following the intial condition. This does technically allow the model to
train on the initial condition point, but that's only 1 data point about
of 61 and the bias should make our result conservative. If we want to
improve this in the future we'll need a particle filter instead.